Create schema for Yup validation
